backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography>=42.0.0
dnspython==2.5.0
stripe==7.12.0
apscheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.3
pandas>=2.1.0
playwright>=1.58.0
aiofiles==23.2.1
python-dotenv==1.0.0
